Fixed interpolation of z at river points. Had _user where it should have been _grid.
Error checked for head_defined in calculation of z_grid from z_user. Head is always defined because it was interpolated earlier in the routine.

Old version had a mistake because of the check for head_defined in tidy_river, but not in write.c. Used value of zero for depth in some cases, which resulted in erroneous Z for nodes. 

git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@3246 1feff8c3-07ed-0310-ac33-dd36852eb9cd