Squashed 'src/phast/PhreeqcRM/' changes from 9cfb5762..80bba0b6
80bba0b6 Merge commit '1f1a092d49ddb56473a97023abc859d751612583'
1f1a092d Squashed 'src/' changes from 2c5fcbaa..8fb64150
6f4503cf Enabled more jobs for testing
6a92a860 Merge pull request #162 from dlparkhurst/module
1a57b074 Merge branch 'master' into module
3c7ffa13 Merge branch 'master' into module
d5d7d751 Issue 160 in phreeqc. llnl minimum temperature problem. There are several places where a SOLUTION is constructed to get lists or test the database. By default the SOLUTION is 25C, so if llnl has a minimum temperature greater (or maximum temperature less) than 25, an error is encountered. The fix is to change the temperature of the SOLUTIONs to the minimum temperature of the llnl grid if llnl is being used.

git-subtree-dir: src/phast/PhreeqcRM
git-subtree-split: 80bba0b69e8053addb8c97ca8139730a2e61db64